+++ packages/expo-modules-core/src/sweet/NativeJSLogger.ts
@@ -1,0 +1,3 @@
+import { requireOptionalNativeModule } from '../requireNativeModule';
+
+export default requireOptionalNativeModule('ExpoModulesCoreJSLogger');